# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IEP — Employee Idea Evaluation & Innovation Platform",
    "version": "1.0.0",
    "description": "GENERATED FILE — do not edit.\n\nSource of truth: packages/contracts/src/api.ts. Regenerate with `pnpm openapi:generate`.\nCI fails if this file is stale (SPEC §11.3). Breaking changes require a superseding ADR (SPEC §14.1)."
  },
  "servers": [
    {
      "url": "/api",
      "description": "IEP API"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-iep.sid"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_FAILED",
              "MALFORMED_REQUEST",
              "UNSUPPORTED_FILE_TYPE",
              "FILE_TOO_LARGE",
              "UNAUTHENTICATED",
              "SESSION_EXPIRED",
              "FORBIDDEN",
              "ROLE_NOT_PERMITTED",
              "NOT_SUBMITTER",
              "CANNOT_REVIEW_OWN_IDEA",
              "NOT_FOUND",
              "ILLEGAL_STATUS_TRANSITION",
              "REASON_REQUIRED",
              "IDEA_VERSION_IMMUTABLE",
              "PROFILE_WEIGHTS_UNBALANCED",
              "CONCURRENT_MODIFICATION",
              "EVIDENCE_REQUIRED",
              "MITIGATION_REQUIRED",
              "CONSTRAINT_CITATION_REQUIRED",
              "EXPLANATION_REQUIRED",
              "RECOMMENDATION_INCOMPLETE",
              "RATE_LIMITED",
              "AI_BUDGET_EXCEEDED",
              "INTERNAL_ERROR",
              "AI_PROVIDER_UNAVAILABLE",
              "AI_OUTPUT_INVALID",
              "DEPENDENCY_UNAVAILABLE",
              "NOT_IMPLEMENTED_UNTIL_M2"
            ]
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "array",
                  "items": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "number"
                      }
                    ]
                  }
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "additionalProperties": false
      },
...
      "GetIdeaResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "SUBMITTED",
              "AI_ANALYSIS",
              "NEEDS_CLARIFICATION",
              "EVALUATED",
              "RANKED",
              "UNDER_REVIEW",
              "PROTOTYPE_CANDIDATE",
              "PILOT",
              "PRODUCTION_CANDIDATE",
              "IMPLEMENTED",
              "PARKED",
              "BLOCKED",
              "REJECTED",
              "ARCHIVED"
            ]
          },
          "maturityLevel": {
            "anyOf": [
              {
                "type": "number",
                "enum": [
                  1
                ]
              },
              {
                "type": "number",
                "enum": [
                  2
                ]
              },
              {
                "type": "number",
                "enum": [
                  3
                ]
              },
              {
                "type": "number",
                "enum": [
                  4
                ]
              },
              {
                "type": "number",
                "enum": [
                  5
                ]
              }
            ],
            "nullable": true
          },
          "submitter": {
            "type": "object",
            "properties": {
              "id": {
                "$ref": "#/components/schemas/definitions/GetIdeaResponse/properties/id"
              },
              "displayName": {
                "type": "string"
              },
              "departmentName": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "id",
              "displayName",
              "departmentName"
            ],
            "additionalProperties": false
          },
          "department": {
            "type": "object",
            "properties": {
              "id": {
                "$ref": "#/components/schemas/definitions/GetIdeaResponse/properties/id"
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name"
            ],
            "additionalProperties": false,
            "nullable": true
          },
          "category": {
            "type": "object",
            "properties": {
              "id": {
                "$ref": "#/components/schemas/definitions/GetIdeaResponse/properties/id"
              },
              "key": {
                "type": "string"
              },
              "label": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "key",
              "label"
            ],
            "additionalProperties": false,
            "nullable": true
          },
          "currentVersionNo": {
            "type": "integer",
            "minimum": 1
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedAt": {
            "$ref": "#/components/schemas/definitions/GetIdeaResponse/properties/submittedAt"
          },
          "rank": {
            "type": "integer",
            "minimum": 1,
            "nullable": true
          },
          "compositeScore": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "nullable": true
          },
          "currentVersion": {
            "type": "object",
            "properties": {
              "id": {
                "$ref": "#/components/schemas/definitions/GetIdeaResponse/properties/id"
              },
              "versionNo": {
                "type": "integer",
                "minimum": 1
              },
              "title": {
                "$ref": "#/components/schemas/definitions/GetIdeaResponse/properties/title"
              },
              "changeSummary": {
                "type": "string",
                "nullable": true
              },
              "author": {
                "$ref": "#/components/schemas/definitions/GetIdeaResponse/properties/submitter"
              },
              "createdAt": {
                "$ref": "#/components/schemas/definitions/GetIdeaResponse/properties/submittedAt"
              },
              "description": {
                "type": "string"
              },
              "problemStatement": {
                "type": "string"
              },
              "expectedUsers": {
                "type": "string"
              },
              "expectedOutcome": {
                "type": "string"
              },
              "existingProcess": {
                "type": "string",
                "nullable": true
              },
              "existingSolutions": {
                "type": "string",
                "nullable": true
              },
              "suggestedTechnology": {
                "type": "string",
                "nullable": true
              },
              "expectedBenefits": {
                "type": "string",
                "nullable": true
              },
              "estimatedCostNote": {
                "type": "string",
                "nullable": true
              },
              "references": {
                "type": "string",
                "nullable": true
              },
              "attachments": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "$ref": "#/components/schemas/definitions/GetIdeaResponse/properties/id"
                    },
                    "filename": {
                      "type": "string"
                    },
                    "mime": {
                      "type": "string"
                    },
                    "bytes": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id",
                    "filename",
                    "mime",
                    "bytes"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "required": [
              "id",
              "versionNo",
              "title",
              "changeSummary",
              "author",
              "createdAt",
              "description",
              "problemStatement",
              "expectedUsers",
              "expectedOutcome",
              "existingProcess",
              "existingSolutions",
              "suggestedTechnology",
              "expectedBenefits",
              "estimatedCostNote",
              "references",
              "attachments"
            ],
            "additionalProperties": false
          },
          "versionCount": {
            "type": "integer",
            "minimum": 1
          },
          "openRecommendationCount": {
            "type": "integer",
            "minimum": 0
          },
          "permissions": {
            "type": "object",
            "properties": {
              "canEdit": {
                "type": "boolean"
              },
              "canSubmit": {
                "type": "boolean"
              },
              "canRevise": {
                "type": "boolean"
              },
              "canReview": {
                "type": "boolean"
              },
              "canOverrideScores": {
                "type": "boolean"
              },
              "allowedTransitions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/definitions/GetIdeaResponse/properties/status"
                }
              }
            },
            "required": [
              "canEdit",
              "canSubmit",
              "canRevise",
              "canReview",
              "canOverrideScores",
              "allowedTransitions"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "maturityLevel",
          "submitter",
          "department",
          "category",
          "currentVersionNo",
          "submittedAt",
          "updatedAt",
          "rank",
          "compositeScore",
          "currentVersion",
          "versionCount",
          "openRecommendationCount",
          "permissions"
        ],
        "additionalProperties": false
      },
...
      "UpdateDraftRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20000
          },
          "problemStatement": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "expectedUsers": {
            "$ref": "#/components/schemas/definitions/UpdateDraftRequest/properties/problemStatement"
          },
          "expectedOutcome": {
            "$ref": "#/components/schemas/definitions/UpdateDraftRequest/properties/problemStatement"
          },
          "existingProcess": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          },
          "existingSolutions": {
            "$ref": "#/components/schemas/definitions/UpdateDraftRequest/properties/existingProcess"
          },
          "suggestedTechnology": {
            "$ref": "#/components/schemas/definitions/UpdateDraftRequest/properties/existingProcess"
          },
          "expectedBenefits": {
            "$ref": "#/components/schemas/definitions/UpdateDraftRequest/properties/existingProcess"
          },
          "estimatedCostNote": {
            "$ref": "#/components/schemas/definitions/UpdateDraftRequest/properties/existingProcess"
          },
          "references": {
            "$ref": "#/components/schemas/definitions/UpdateDraftRequest/properties/existingProcess"
          },
          "departmentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "categoryId": {
            "allOf": [
              {
                "$ref": "#/components/schemas/definitions/UpdateDraftRequest/properties/departmentId"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/ideas/{ideaId}": {
      "get": {
        "operationId": "getIdea",
        "summary": "Idea detail, including what this actor is permitted to do with it.",
        "tags": [
          "ideas"
        ],
        "x-access": {
          "requires": [
            "idea:read:own"
          ]
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "ideaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetIdeaResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateDraft",
        "summary": "Edit a DRAFT or NEEDS_CLARIFICATION idea in place. Submitted ideas are immutable.",
        "tags": [
          "ideas"
        ],
        "x-access": {
          "requires": [
            "idea:edit:own"
          ]
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "ideaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateDraftResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "IDEA_VERSION_IMMUTABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDraftRequest"
              }
            }
          }
        }
      }
    },
...
  }
}